+++ docs/changelog.rst
@@ -8,6 +8,8 @@ v0.6.0 (in development)
 - Gave :program:`javaproperties`' :program:`format` and :program:`select`
   subcommands ``--ascii`` and ``--unicode`` options for controlling the
   escaping of non-ASCII characters in output
+- Gave :program:`properties2json` ``--ascii`` and ``--unicode`` options for
+  controlling the escaping of non-ASCII characters in output
 
 v0.5.0 (2020-01-24)
 -----------------------